Move out virtualServiceRegistry from workspace folder
