+++ packages/serverless-logic-sandbox/src/virtualServiceRegistry/models/BaseOpenApiSpec.ts
